--- backend/pokerogue/src/main/java/com/pokerogue/external/dto/DataUrl.java
@@ -1,4 +1,0 @@
-package com.pokerogue.external.dto;
-
-public record DataUrl(String name, String url) {
-}
