--- backend/pokerogue/src/main/java/com/pokerogue/external/dto/Language.java
@@ -1,8 +1,0 @@
-package com.pokerogue.external.dto;
-
-public record Language(String name) {
-
-    public boolean isKorean() {
-        return name.equals("ko");
-    }
-}
