--- backend/pokerogue/src/main/java/com/pokerogue/external/dto/Name.java
@@ -1,8 +1,0 @@
-package com.pokerogue.external.dto;
-
-public record Name(Language language, String name) {
-
-    public boolean isKorean() {
-        return language.isKorean();
-    }
-}
