--- backend/pokerogue/src/main/java/com/pokerogue/external/dto/ability/FlavorTextEntry.java
@@ -1,6 +1,0 @@
-package com.pokerogue.external.dto.ability;
-
-import com.pokerogue.external.dto.Language;
-
-public record FlavorTextEntry(String flavor_text, Language language) {
-}
