--- backend/pokerogue/src/main/java/com/pokerogue/external/dto/ability/PokemonSummary.java
@@ -1,6 +1,0 @@
-package com.pokerogue.external.dto.ability;
-
-import com.pokerogue.external.dto.DataUrl;
-
-public record PokemonSummary(DataUrl pokemon) {
-}
