--- backend/pokerogue/src/main/java/com/pokerogue/external/dto/pokemon/AbilityDataUrl.java
@@ -1,6 +1,0 @@
-package com.pokerogue.external.dto.pokemon;
-
-import com.pokerogue.external.dto.DataUrl;
-
-public record AbilityDataUrl(DataUrl ability) {
-}
