--- backend/pokerogue/src/main/java/com/pokerogue/external/dto/pokemon/StatDetail.java
@@ -1,10 +1,0 @@
-package com.pokerogue.external.dto.pokemon;
-
-import com.pokerogue.external.dto.DataUrl;
-
-public record StatDetail(int base_stat, DataUrl stat) {
-
-    public String getStatName() {
-        return stat.name();
-    }
-}
