--- backend/pokerogue/src/main/java/com/pokerogue/external/dto/pokemon/species/PokemonNameAndDexNumber.java
@@ -1,4 +1,0 @@
-package com.pokerogue.external.dto.pokemon.species;
-
-public record PokemonNameAndDexNumber(Long pokedexNumber, String koName) {
-}
