+++ backend/pokerogue/src/main/java/com/pokerogue/external/pokemon/dto/DataUrl.java
@@ -1,0 +1,4 @@
+package com.pokerogue.external.pokemon.dto;
+
+public record DataUrl(String name, String url) {
+}
