+++ backend/pokerogue/src/main/java/com/pokerogue/external/pokemon/dto/Language.java
@@ -1,0 +1,8 @@
+package com.pokerogue.external.pokemon.dto;
+
+public record Language(String name) {
+
+    public boolean isKorean() {
+        return name.equals("ko");
+    }
+}
