+++ backend/pokerogue/src/main/java/com/pokerogue/external/pokemon/dto/Name.java
@@ -1,0 +1,8 @@
+package com.pokerogue.external.pokemon.dto;
+
+public record Name(Language language, String name) {
+
+    public boolean isKorean() {
+        return language.isKorean();
+    }
+}
