+++ backend/pokerogue/src/main/java/com/pokerogue/external/pokemon/dto/ability/FlavorTextEntry.java
@@ -1,0 +1,6 @@
+package com.pokerogue.external.pokemon.dto.ability;
+
+import com.pokerogue.external.pokemon.dto.Language;
+
+public record FlavorTextEntry(String flavor_text, Language language) {
+}
