+++ backend/pokerogue/src/main/java/com/pokerogue/external/pokemon/dto/ability/PokemonSummary.java
@@ -1,0 +1,6 @@
+package com.pokerogue.external.pokemon.dto.ability;
+
+import com.pokerogue.external.pokemon.dto.DataUrl;
+
+public record PokemonSummary(DataUrl pokemon) {
+}
