+++ backend/pokerogue/src/main/java/com/pokerogue/external/pokemon/dto/pokemon/AbilityDataUrl.java
@@ -1,0 +1,6 @@
+package com.pokerogue.external.pokemon.dto.pokemon;
+
+import com.pokerogue.external.pokemon.dto.DataUrl;
+
+public record AbilityDataUrl(DataUrl ability) {
+}
