+++ backend/pokerogue/src/main/java/com/pokerogue/external/pokemon/dto/pokemon/StatDetail.java
@@ -1,0 +1,10 @@
+package com.pokerogue.external.pokemon.dto.pokemon;
+
+import com.pokerogue.external.pokemon.dto.DataUrl;
+
+public record StatDetail(int base_stat, DataUrl stat) {
+
+    public String getStatName() {
+        return stat.name();
+    }
+}
