+++ backend/pokerogue/src/main/java/com/pokerogue/external/pokemon/dto/pokemon/species/PokemonNameAndDexNumber.java
@@ -1,0 +1,4 @@
+package com.pokerogue.external.pokemon.dto.pokemon.species;
+
+public record PokemonNameAndDexNumber(Long pokedexNumber, String koName) {
+}
